Analysis Of IPL 2022
Board of Control for Cricket in India (BCCI) is set to earn a whopping sum of a little over Rs 800 crore (INR 8 billion) from the sponsorship and other deals for the Indian Premier League (IPL) 2022. Interesting to analyse this business model, through the lens of basics with matches and player analysis.